requirements.txt
django==3.1
psycopg2==2.8.6
gunicorn==20.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
